+++ magda-web-client/src/Components/RequestDataset/RequestFormLogic.js
@@ -47,7 +47,9 @@ export default class RequestFormLogic extends React.Component {
         } else {
             url =
                 config.correspondenceApiUrl +
-                `send/dataset/${this.props.datasetId}/report`;
+                `send/dataset/${encodeURIComponent(
+                    this.props.datasetId
+                )}/question`;
         }
         fetch(url, {
             method: "POST",